refactor(release-service): split control API layer
