+++ apps/release-service/src/api/errors.ts
@@ -11,7 +11,6 @@ export type ApiErrorCode =
 	| "CREDENTIAL_REVOKED"
 	| "CSRF_INVALID"
 	| "IDEMPOTENCY_KEY_INVALID"
-	| "IDEMPOTENCY_CONFLICT"
 	| "INTERNAL_ERROR"
 	| "INVALID_REQUEST"
 	| "INTENT_NOT_APPROVABLE"
